Improve deserialization typing and fix some bugs
There was an issue when deserializing a `mach_header_64` where `cpusubtype` (and another field I think) were being deserialized to the wrong integer type and causing an assertion failure. I've added a macro to avoid these simple issues in the future.

Additionally there was another bug with deserializing `std::pair<uint64_t, std::pair<uint64_t, uint64_t>>` where it was not being serialized in an expected way. The deserialization has been corrected and the unused serialization function for it has been removed.